refactor shared patches' package
