+++ src/main/kotlin/app/revanced/patches/shared/misc/fix/spoof/annotations/ClientSpoofCompatibility.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.shared.fix.spoof.annotations
+package app.revanced.patches.shared.misc.fix.spoof.annotations
 
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
