+++ src/main/kotlin/app/revanced/patches/shared/misc/fix/spoof/fingerprints/UserAgentHeaderBuilderFingerprint.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.shared.fix.spoof.fingerprints
+package app.revanced.patches.shared.misc.fix.spoof.fingerprints
 
 
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
