+++ src/main/kotlin/app/revanced/patches/shared/misc/fix/verticalscroll/patch/VerticalScrollPatch.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.shared.fix.verticalscroll.patch
+package app.revanced.patches.shared.misc.fix.verticalscroll.patch
 
 import app.revanced.extensions.toErrorResult
 import app.revanced.patcher.annotation.Description
@@ -9,8 +9,8 @@ import app.revanced.patcher.extensions.instruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchResult
 import app.revanced.patcher.patch.PatchResultSuccess
-import app.revanced.patches.shared.fix.verticalscroll.annotations.VerticalScrollCompatibility
-import app.revanced.patches.shared.fix.verticalscroll.fingerprints.CanScrollVerticallyFingerprint
+import app.revanced.patches.shared.misc.fix.verticalscroll.annotations.VerticalScrollCompatibility
+import app.revanced.patches.shared.misc.fix.verticalscroll.fingerprints.CanScrollVerticallyFingerprint
 import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
 
 @Description("Fixes issues with scrolling on the home screen when the first component is of type EmptyComponent.")
